♻️ move types to on files
