+++ api/src/api/v1/types/general.ts
@@ -1,0 +1,23 @@
+import { ApiRequestQuery, ApiRequestBody } from '../schema/request';
+import { ApiResponse } from '../schema/response';
+
+/*
+ * api.json related types
+ */
+export enum HttpMethodEnum {
+  GET = 'GET',
+  POST = 'POST',
+  PUT = 'PUT',
+  DELETE = 'DELETE',
+}
+
+export type ApiRouteSpec = {
+  description: string,
+  isImplemented: boolean,
+  auth: boolean,
+  intendedFor: any,
+  scope: string[],
+  query: ApiRequestQuery,
+  body: ApiRequestBody,
+  response: ApiResponse,
+};
