+++ api/src/api/v1/types/prereqs.ts
@@ -1,0 +1,10 @@
+import * as Hapi from '@hapi/hapi';
+
+/*
+ * Prereq types
+ */
+export interface RequireSiblingPreReq extends Hapi.Request {
+  params: {
+    userId: string
+  };
+}
